gen: Added --install (-i) and --devInstall (-d) options
Both can be specified multiple times and accept single or "space separated" list of pakage names.

  --install behaves like 'npm install --save'
  --devInstall behaves like 'npm install --save-dev'

Package names may include a '@version' suffix even with caret or tilde ranges.

If @version is not specified, current @latest tag is used if possible, otherwise literal '@latest' is used.

If @version is specified and is not a caret or tilde range, it is checked against npm register and a warning is emitted if not exists or cannot be verified. But it is used anyway.